requirements.txt
cantera
numpy
matplotlib
pytest
pyyaml
scipy
jupyter
